Changed the point of the Node Server. Now a WebSocket Server.
