+++ README.md
@@ -1,2 +1,2 @@
 # node-polyglot
-Rails server to run alongside https://github.com/beatriceo/polyglot
+Node server to run alongside https://github.com/beatriceo/polyglot
